requirements.txt
anthropic>=0.40.0
python-docx>=1.1.0
pdfplumber>=0.11.0
python-pptx>=1.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
